feat: move reading file to separate hook
